chore: update enpoint to fetch state with subgraph fallback
