+++ src/app/api/db/constants.ts
@@ -61,6 +61,23 @@ export const BTC_VAULT_NAV_HISTORY_COLUMNS = [
   'transactionHash',
 ]
 
+export const BTC_VAULT_AUDIT_LOG_COLUMNS = [
+  COLUMN_ID,
+  'vault',
+  'type',
+  'amountInWei',
+  'detail',
+  'isNative',
+  'role',
+  'actor',
+  'from',
+  'destination',
+  'blockNumber',
+  'blockTimestamp',
+  'transactionHash',
+  'logIndex',
+]
+
 const BACKER_TO_BUILDER_BACKER = 'backer'
 const BACKER_TO_BUILDER_BUILDER = 'builder'
 const BACKER_TO_BUILDER_TOTAL_ALLOCATION = 'totalAllocation'
